converted particle_t to use floats

+++ soar-robotics/splinter-soar/src/splintersoar/pf/ParticleFilter.java
@@ -56,14 +56,14 @@ public class ParticleFilter {
 
 		particleslcm = new particles_t();
 		particleslcm.nparticles = size;
-		particleslcm.particle = new double[size][];
+		particleslcm.particle = new float[size][];
 
 		for (int i = 0; i < 100; i++) {
 			Particle p = new Particle();
 			p.xyt = new double[] { (1 - 2 * rand.nextFloat()) * 5, (1 - 2 * rand.nextFloat()) * 5, 2 * rand.nextFloat() * Math.PI };
 			particles.add(p);
 
-			particleslcm.particle[i] = new double[] { p.xyt[0], p.xyt[1], p.xyt[2] };
+			particleslcm.particle[i] = new float[] { (float)p.xyt[0], (float)p.xyt[1], (float)p.xyt[2] };
 			
 			if (logger.isLoggable(Level.FINEST)) {
 				logger.finest(String.format("New particle xyt: %5.3f %5.3f %5.3f", p.xyt[0], p.xyt[1], p.xyt[2]));
@@ -142,7 +142,7 @@ public class ParticleFilter {
 
 		particleslcm = new particles_t();
 		particleslcm.nparticles = particles.size();
-		particleslcm.particle = new double[particles.size()][];
+		particleslcm.particle = new float[particles.size()][];
 
 		for (int i = 0; i < particles.size(); i++) {
 			double tw = rand.nextFloat();
@@ -162,7 +162,7 @@ public class ParticleFilter {
 			np.xyt = new double[] { p.xyt[0] + rand.nextGaussian() * 0.05, p.xyt[1] + rand.nextGaussian() * 0.05, p.xyt[2] + rand.nextGaussian() * 0.01 };
 			newParticles.add(np);
 
-			particleslcm.particle[i] = new double[] { np.xyt[0], np.xyt[1], np.xyt[2] };
+			particleslcm.particle[i] = new float[] { (float)np.xyt[0], (float)np.xyt[1], (float)np.xyt[2] };
 			
 			if (logger.isLoggable(Level.FINEST)) {
 				logger.finest(String.format("Resampled particle xyt: %5.3f %5.3f %5.3f", np.xyt[0], np.xyt[1], np.xyt[2]));
